Add asterisks, description, checkbox
Add asterisks SVGs to 'Title', 'Author', and 'Page Count' to denote
required input fields. Change 'Page Count' input type to "number" so that
it will only accept numbers.

Add optional 'Description' field as well as 'Read' checkbox so the reader
can denote if they have read the title or not.