Refactor displayBooks function
Place code in neatly organized groups for better legibility. Refactor
iteration through myLibrary array. Originally, there was no 'if' statement
to provide condition for the array loop. The array would loop through in full,
each time the 'submit' button was pressed. This caused duplication of the divs
in the DOM. Fix with an if statement.