Add if statment exceptions
Add 'if' statements to displayBooks to account for edge cases. Since
default objects were not physically checked, add "Check" to the if/statement
to show that they've been read upon page load.

Add another condition for the description element. If no description is
added, 'li' element will not even be created.